// x-pack/platform/plugins/shared/fleet/server/types/so_attributes.d.ts
export interface CloudConnectorSOAttributes {
    name: string;
    namespace?: string;
    cloudProvider: CloudProvider;
    accountType?: AccountType;
    vars: CloudConnectorVars;
    created_at: string;
    updated_at: string;
    verification_status?: VerificationStatus;
    verification_started_at?: string;
    verification_failed_at?: string;
}